+++ lib/pages/login/login_form_data.dart
@@ -1,0 +1,6 @@
+class LoginFormData {
+  String username;
+  String password;
+
+  LoginFormData();
+}
