+++ lib/pages/register/new_user_form_data.dart
@@ -1,0 +1,7 @@
+class NewUserFormData {
+  String username;
+  String name;
+  String password;
+
+  NewUserFormData();
+}
